Adding the actual GAN being used to the repository.
I am slowly modifying the code to suit my needs for this project. It
has diverged enough from the original that I would need to include
large portions of it anyways, so I will instead include the entire
thing.

All credit for original code is kept in the Attributions file, and
a permissive license has been chosen for the project.