Update Network to work with 512x512 files.
Squash Commit.

This commit contains alterations to the network that allow it to work on files
of size 512x512 rather than 256x256.

The amount of changes is larger than I should have done in a single PR, but so
it goes. I will endeavor to be better at segmenting PRs in the future.

The changes include:

- Adding resize layers in the Generator to allow it to work on larger inputs
  without increasing VRAM usage. I assume that the performance is worse than
  natively working with those resolutions, but I do not have the GPU to test
  this.

- Altering the network internals to use a Spatial Pyramid Attentive Pooling
  model (SPAP) instead of a Residual Network (for the Generator) or a PathGAN
  (for the discriminator). See here for information:
  https://arxiv.org/pdf/1901.06322.pdf